SilentLimitedBuffer fix: UnicodeDecodeError
Multibyte UTF8 characters could be in the outputbuffer and they might be
chopped. If they are decode would throw an exception, this prevents
this.
